+++ pom.xml
@@ -3,7 +3,7 @@
   <modelVersion>4.0.0</modelVersion>
 
   <groupId>com.eniacdevelopment.eniachome.datamodel</groupId>
-  <artifactId>EniacHomeDataModel</artifactId>
+  <artifactId>DataModel</artifactId>
   <version>1.0-SNAPSHOT</version>
   <packaging>jar</packaging>
