+++ src/main/java/com/eniacdevelopment/eniachome/DataModel/AlarmStatusPackage.java
@@ -1,4 +1,4 @@
-package com.eniacdevelopment.serialization.DataModel;
+package com.eniacdevelopment.eniachome.DataModel;
 
 /**
  * Created by larsg on 10/13/2016.
