+++ src/main/java/com/eniacdevelopment/eniachome/DataModel/User/CreateUserPackage.java
@@ -1,4 +1,4 @@
-package com.eniacdevelopment.serialization.DataModel.User;
+package com.eniacdevelopment.eniachome.DataModel.User;
 
 /**
  * Created by larsg on 10/13/2016.
